refactor: move agent into internal
