--- agent/.gitignore
@@ -1,36 +1,0 @@
-build
-.devcontainer/shared/*.rpm
-*.sqlite
-tests/cli/*.diff
-tests/cli/*.exp
-tests/cli/*.log
-tests/cli/*.out
-tests/cli/*.php
-tests/cli/*.sh
-tests/cli/**/*.diff
-tests/cli/**/*.exp
-tests/cli/**/*.log
-tests/cli/**/*.out
-tests/cli/**/*.php
-tests/cli/**/*.sh
-tests/cli/**/*.post
-*aikido-agent
-*.pb.go
-*.pb.go
-*.log
-*.nice
-.vscode/settings.json
-*.deb
-lib/agent/aikido_config.json
-tests/testlib/__pycache__/*.pyc
-benchmarks/server/*.txt
-benchmarks/cli/*.txt
-benchmarks/cli/benchmark_results.md
-benchmarks/server/benchmark_results.md
-tools/server_tests/apache/__pycache__/*.pyc
-tools/server_tests/nginx/__pycache__/*.pyc
-tools/server_tests/php_built_in/__pycache__/*.pyc
-tests/server/**/*.conf
-dist*
-# Editors
-.idea*
