+++ internal/agent/api_discovery/get_data_schema.go
@@ -3,7 +3,7 @@ package api_discovery
 import (
 	"reflect"
 
-	"github.com/AikidoSec/firewall-go/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 )
 
 const (
