+++ internal/agent/api_discovery/merge_data_schemas.go
@@ -3,7 +3,7 @@ package api_discovery
 import (
 	"reflect"
 
-	"github.com/AikidoSec/firewall-go/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 )
 
 // MergeDataSchemas merges two DataSchema objects.
