+++ internal/agent/api_discovery/merge_data_schemas_test.go
@@ -3,7 +3,7 @@ package api_discovery
 import (
 	"testing"
 
-	"github.com/AikidoSec/firewall-go/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 
 	"github.com/stretchr/testify/assert"
 )
