+++ internal/agent/cloud/event_heartbeat.go
@@ -3,9 +3,9 @@ package cloud
 import (
 	"sync/atomic"
 
-	. "github.com/AikidoSec/firewall-go/agent/aikido_types"
-	"github.com/AikidoSec/firewall-go/agent/globals"
-	"github.com/AikidoSec/firewall-go/agent/utils"
+	. "github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/globals"
+	"github.com/AikidoSec/firewall-go/internal/agent/utils"
 )
 
 func GetHostnamesAndClear() []Hostname {
