+++ internal/agent/globals/constants.go
