+++ internal/agent/request.go
@@ -1,11 +1,11 @@
 package agent
 
 import (
-	"github.com/AikidoSec/firewall-go/agent/aikido_types"
-	"github.com/AikidoSec/firewall-go/agent/api_discovery"
-	"github.com/AikidoSec/firewall-go/agent/globals"
-	"github.com/AikidoSec/firewall-go/agent/log"
-	"github.com/AikidoSec/firewall-go/agent/utils"
+	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/api_discovery"
+	"github.com/AikidoSec/firewall-go/internal/agent/globals"
+	"github.com/AikidoSec/firewall-go/internal/agent/log"
+	"github.com/AikidoSec/firewall-go/internal/agent/utils"
 )
 
 func storeStats() {
