+++ internal/helpers/match_endpoints_test.go
@@ -3,7 +3,7 @@ package helpers
 import (
 	"testing"
 
-	. "github.com/AikidoSec/firewall-go/agent/aikido_types"
+	. "github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 	"github.com/stretchr/testify/assert"
 )
