+++ internal/http/ip_allowed_to_access_route_test.go
@@ -3,7 +3,7 @@ package http
 import (
 	"testing"
 
-	. "github.com/AikidoSec/firewall-go/agent/aikido_types"
+	. "github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
 	"github.com/stretchr/testify/assert"
 )
