+++ zen/should_block_request.go
@@ -3,7 +3,7 @@ package zen
 import (
 	"context"
 
-	"github.com/AikidoSec/firewall-go/agent"
+	"github.com/AikidoSec/firewall-go/internal/agent"
 	"github.com/AikidoSec/firewall-go/internal/config"
 	"github.com/AikidoSec/firewall-go/internal/helpers"
 	"github.com/AikidoSec/firewall-go/internal/log"
